requirements.txt
tqdm
ninja
numpy
scipy
matplotlib
opencv-python
imageio
ftfy
omegaconf
git+https://github.com/openai/CLIP.git
pytorch-lightning
torchmetrics
huggingface_hub
transformers
collections
numpy 
scipy 
kornia 
pandas 
opencv-python 
pillow
tqdm
ftfy
regex